Change handling of tail in //seq, for better failure result
Previously, //seq created a temporary array for checking tail.  This made
it impossible to produce an appropriate data path in the failure.

I solved this issue by adding a skip option to //arr, which specifies how
many items at the start of the list to skip, for both the length and
contents checks.

//seq now passes the original data array to //arr, with skip set to the
size of //seq's contents array.

Added appropriate tests.