Add data_type and check_type info to failure structs
Indicates whether each item in the data and check info is a hash key ('k)
or an array index ('i').  Allows creating the path strings with the correct
subscript types.  Not a very elegant solution, unfortunately.

(There may be some irony that the structure of parallel arrays can't be
validated with Data::Rx.  I didn't want to change the data and check parts
themselves.)