Pass original value of 'type' to new_checker()
For example, if the schema is { type => '//str' }, then '//str' is passed
as the first argument to new_checker(), even though it was expanded to
tag:codesimply.com,2008:rx/core/str for the lookup.

This will allow error messages to refer to the type as the user entered it,
regardless of standard or non-standard prefix used.